// src/types/ip-address.type.ts
export interface IpAddress {
    status: string;
    query: string;
    continent: string;
    country: string;
    regionName: string;
    city: string;
    timezone: string;
    isp: string;
    org: string;
    proxy: boolean;
    mobile: boolean;
    hosting: boolean;
    lat: number;
    lon: number;
}